Fix decryption instructions when using openssl 1.1
The problem:
openssl 1.0 has been deprecated because it can be hacked and is not
secure. openssl 1.0 has been removed from most Linux repositories. But
version 1.1.0 uses a new digest, so the command on the README doesn't
work.

The instructions on the README would just output "bad decrypto".

How to fix it:
You can force openssl 1.1.0 to use the (old) md5 hash by adding "-md
md5" to the list of params.